+++ 3DS/source/keyboard.c
@@ -2,7 +2,7 @@
 #include <math.h>
 #include <string.h>
 
-#include "font.h"
+#include "drawing.h"
 
 #include "keyboard.h"
