+++ 3DS/source/main.c
@@ -6,7 +6,7 @@
 
 #include "wireless.h"
 #include "settings.h"
-#include "font.h"
+#include "drawing.h"
 #include "input.h"
 #include "keyboard.h"
